+++ src/felupe/view/__init__.py
@@ -1,0 +1,13 @@
+from ._field import ViewField
+from ._mesh import ViewMesh
+from ._scene import Scene
+from ._solid import ViewSolid
+from ._xdmf import ViewXdmf
+
+__all__ = [
+    "Scene",
+    "ViewField",
+    "ViewMesh",
+    "ViewSolid",
+    "ViewXdmf",
+]